lundi3016h

+++ ft_strncmp.c
@@ -11,19 +11,20 @@
 /* ************************************************************************** */
 
 /*#include <stdio.h>*/
+#include "libc.h"
 
 int ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-    int i;
+	int i;
 
-    i = 0;
-    while (i < n)
-    {
-        if (s1[i] != s2[i])
-            return (s1[i] - s2[i]);
-        i++;
-    }
-    return (0);
+	i = 0;
+	while (i < (int)n)
+	{
+		if (s1[i] != s2[i])
+			return (s1[i] - s2[i]);
+		i++;
+	}
+	return (0);
 }
 
 /*int main(void)
